requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
filelock==3.16.1
mysqlclient==2.2.5
platformdirs==4.3.6
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.27.0
